+++ stepper.d
@@ -2,12 +2,12 @@ import std.math;
 import coal_state;
 
 class Stepper {
-    size_t nr_steps;
+    int nr_steps;
     double tMax;
     double alpha;
     double[] time_boundaries;
     
-    this(size_t nr_steps=10000, double tMax=20, double alpha=0.001) {
+    this(int nr_steps=10000, double tMax=20, double alpha=0.001) {
         this.nr_steps = nr_steps;
         this.alpha = alpha;
         this.tMax = tMax;
@@ -16,7 +16,7 @@ class Stepper {
             time_boundaries[i] = time_function(i);
     }
     
-    double time_function(size_t i) const {
+    double time_function(int i) const {
         return alpha * exp(cast(double)(i) / nr_steps * log(1.0 + tMax / alpha)) - alpha;
     }
     
@@ -30,7 +30,7 @@ class Stepper {
 
 unittest {
     import model;
-    auto nVec = [500UL, 500, 500];
+    auto nVec = [500, 500, 500];
     auto joins = [Join_t(0.1, 0, 1, 1.0), Join_t(0.2, 0, 2, 1.0)];
     auto m = new Model(nVec, [1.0, 1.0, 1.0], joins);
